Crash in WebCore::RenderTableSection::rowLogicalHeightChanged https://webkit.org/b/76842
Reviewed by Darin Adler.

Source/WebCore:

Test: fast/table/crash-section-logical-height-changed-needsCellRecalc.html

The issue was that we would access our section's structure when it was dirty.

* rendering/RenderTableSection.cpp:
(WebCore::RenderTableSection::rowLogicalHeightChanged):
Bail out if we need cells recalculation as our internal structure is not up-to-date
and we will recompute all the rows' heights as part of the recomputation anyway.

LayoutTests:

* fast/table/crash-section-logical-height-changed-needsCellRecalc-expected.txt: Added.
* fast/table/crash-section-logical-height-changed-needsCellRecalc.html: Added.


Identifier: 93718@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@105685 268f45cc-cd09-0410-ab3c-d52691b4dbfc