In CodeGeneratorV8.pm, overwrite the output .h/.cpp only if the bytes differ https://bugs.webkit.org/show_bug.cgi?id=76920
Reviewed by Darin Adler.

This is one of steps to stop rebuilding .h/.cpp files
generated by unchanged IDLs (bug 76836).
This patch makes a change on CodeGeneratorV8.pm so that
it overwrites the output .h/.cpp only if the bytes differ.

No tests. No change in behavior.
I manually confirmed that when I add a new attribute to Element.idl,
the time-stamps of unrelated V8*.h and V8*.cpp do not change.

* bindings/scripts/CodeGeneratorV8.pm:
(WriteData): Used UpdateFileIfChanged().


Identifier: 93796@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@105766 268f45cc-cd09-0410-ab3c-d52691b4dbfc