CodeGeneratorCPP.pm should overwrite the output .h/.cpp only if the bytes differ https://bugs.webkit.org/show_bug.cgi?id=76926
Reviewed by Adam Barth.

This is one of steps to stop rebuilding .h/.cpp files
generated by unchanged IDLs (bug 76836).
This patch makes a change on CodeGeneratorCPP.pm so that
it overwrites the output .h/.cpp only if the bytes differ.

No tests. No change in behavior.
Manually confirm that when you add a new attribute to Element.idl,
the time-stamps of unrelated WebDOM*.h and WebDOM*.cpp do not change.

* bindings/scripts/CodeGeneratorCPP.pm:
(WriteData): Used UpdateFileIfChanged().


Identifier: 93819@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@105805 268f45cc-cd09-0410-ab3c-d52691b4dbfc