[chromium] Fix css3/flexbox/cross-axis-scrollbar-expected to work on chromium win.
The scrollbars overflow the bounds of the scrollbar rect (see bug 77368).
Work around this by using positioned elements and setting the z-index as
negative.  It didn't work previously because when the scrollbar was
positioned, it clipped the extra pixels.  Tested to verify that this works
on Chromium Win and Chromium Linux.

Unreviewed, test fix.

* css3/flexbox/cross-axis-scrollbar-expected.html:
* platform/chromium/test_expectations.txt:

Identifier: 94332@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@106378 268f45cc-cd09-0410-ab3c-d52691b4dbfc