[Chromium] Expose setLayoutFallbackWidth as a WebSetting https://bugs.webkit.org/show_bug.cgi?id=77431
Reviewed by Darin Fisher.

The layout fallback width is the default fixed layout width
set for desktop web pages that lack a viewport tag. This width
can vary from platform to platform and can depend on other factors,
and so it should be a setting exposed to Chromium.

* public/WebSettings.h:
():
* src/WebSettingsImpl.cpp:
(WebKit::WebSettingsImpl::setLayoutFallbackWidth):
(WebKit):
* src/WebSettingsImpl.h:
(WebSettingsImpl):


Identifier: 94348@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@106394 268f45cc-cd09-0410-ab3c-d52691b4dbfc