[skia] Fix unit test PlatformContextSkiaTest.trackOpaqueOvalTest https://bugs.webkit.org/show_bug.cgi?id=77738
We need to clear the bitmap before the test runs, otherwise the
contents of the bitmap will be uninitialised. This causes an
unpredicatable result as the canvas draw operations blend the
destination bitmap with the source colour.

Similarly, valgrind discovered that
PlatformContextSkiaTest.trackOpaqueTest and
PlatformContextSkiaTest.trackOpaqueLineTest suffers from the same
problem.

Also, re-enable WebFrameTest.FindInPage which was falsely
blamed for making this test fail.

Reviewed by Stephen White.

* tests/PlatformContextSkiaTest.cpp:
(WebCore::TEST):
* tests/WebFrameTest.cpp:
(WebKit::TEST_F):


Identifier: 94609@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@106669 268f45cc-cd09-0410-ab3c-d52691b4dbfc