Unreviewed chromium test expectation change
Unmarking storage/indexeddb/data-corruption.html as TIMEOUT. Was apparently
bulk-marked as TIMEOUT on Snow Leopard as part of WK54322 although it is
not mentioned in that issue. The test hasn't failed in recent history and
there has been substantial rework on the IndexedDB feature since the
expectation was set.

* platform/chromium/test_expectations.txt:

Identifier: 94766@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@106842 268f45cc-cd09-0410-ab3c-d52691b4dbfc