http/tests/security/xssAuditor/cookie-injection.html is flaky https://bugs.webkit.org/show_bug.cgi?id=77908
Reviewed by Eric Seidel.

Before this patch, the output of this test depended on the state of the
global cookie store.  When running tests in parallel, the cookie store
is shared by multiple tests, and so its state varies from run to run.

This patch changes this test to look only for the specific cookie used
by this test, which should make the test more reliable when run in
parallel with other tests.

* http/tests/security/xssAuditor/cookie-injection-expected.txt:
* http/tests/security/xssAuditor/resources/echo-intertag.pl:
* platform/chromium/test_expectations.txt:


Identifier: 94824@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@106906 268f45cc-cd09-0410-ab3c-d52691b4dbfc