[chromium] Add a layout test for losing the compositor context with a video playing https://bugs.webkit.org/show_bug.cgi?id=78060
Reviewed by Eric Carlson.

Adds a test to prevent regressions with video playback when we lose the compositor context. This test loads up a
video, forces a frame to be displayed, simulates a lost context, then finishes the test (which causes another
frame to be displayed).

Prior to r106610 this test times out in debug and crashes in ASAN.

* platform/chromium/compositing/lost-compositor-context-with-video-expected.png: Added.
* platform/chromium/compositing/lost-compositor-context-with-video-expected.txt: Added.
* platform/chromium/compositing/lost-compositor-context-with-video.html: Added.

Identifier: 95043@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@107132 268f45cc-cd09-0410-ab3c-d52691b4dbfc