Improve http-equiv content-language parsing https://bugs.webkit.org/show_bug.cgi?id=77724
Patch by Matt Falkenhagen <falken@chromium.org> on 2012-02-09
Reviewed by Alexey Proskuryakov.

Add tests for mapping http-equiv content-language to -webkit-locale to cover dynamic changes to the meta
element and a comma-separated list of languages. The desired behavior of some of these is not yet decided, so
some of the tests currently fail and the expectations may change. See comments in bug and in the tests.

* fast/text/content-language-comma-separated-list-expected.txt: Added.
* fast/text/content-language-comma-separated-list.html: Added.
* fast/text/content-language-dynamically-added-expected.txt: Added.
* fast/text/content-language-dynamically-added.html: Added.
* fast/text/content-language-dynamically-changed-expected.txt: Added.
* fast/text/content-language-dynamically-changed.html: Added.
* fast/text/content-language-dynamically-removed-expected.txt: Added.
* fast/text/content-language-dynamically-removed.html: Added.
* fast/text/content-language-multiple.html: Fix to actually test http-equiv content-language.

Identifier: 95107@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@107200 268f45cc-cd09-0410-ab3c-d52691b4dbfc