NRWT is broken for the default port https://bugs.webkit.org/show_bug.cgi?id=78640
Reviewed by Tony Chang.

The change in r107718 (bug 78565) broke NRWT if you didn't
specify any sort of --platform flag or equivalent. I didn't
catch this because it's a full integration test (which we don't
have an automated test for) and I missed this command line
variant when testing interactively.

* Scripts/webkitpy/layout_tests/controllers/worker.py:
(Worker.run):

Identifier: 95590@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@107746 268f45cc-cd09-0410-ab3c-d52691b4dbfc